+++ solutions/uri/1011/1011.c
@@ -3,7 +3,7 @@
 
 #define pi 3.14159
 
-int16_t main() {
+int main() {
     float r;
 
     while (scanf("%f", &r) != EOF) {
